feat(issue6): implementa trocas de contexto e métricas obrigatórias #6

+++ include/process.h
@@ -7,31 +7,31 @@
 typedef enum { PROCESS_NEW, PROCESS_READY, PROCESS_RUNNING, PROCESS_BLOCKED, PROCESS_FINISHED } ProcessState;
 
 typedef struct Burst {
-    int cpu_time;
-    int io_time;       // 0 se for a última rajada
+    int64_t cpu_time;
+    int64_t io_time;       // 0 se for a última rajada
     struct Burst *next;
 } Burst;
 
 typedef struct Process {
     int pid;
-    int arrival_time;
+    int64_t arrival_time;
     int priority;
     Burst *bursts;           // lista ligada de rajadas
     Burst *current_burst;    // posição atual
-    int remaining_cpu;       // tempo restante na rajada atual
-    int io_finish_time;      // quando a E/S termina
+    int64_t remaining_cpu;       // tempo restante na rajada atual
+    int64_t io_finish_time;      // quando a E/S termina
     ProcessState state;
     // Acumuladores para métricas
-    int total_cpu_original;
-    int total_io_original;
-    int start_time;          // para turnaround
-    int finish_time;         // para turnaround
+    int64_t total_cpu_original;
+    int64_t total_io_original;
+    int64_t start_time;
+    int64_t finish_time;
     
-    int ready_since;         // instante de entrada atual na fila de prontos
+    int64_t ready_since;         // instante de entrada atual na fila de prontos
 } Process;
 
-Process* process_create(int pid, int arrival_time, int priority);
+Process* process_create(int pid, int64_t arrival_time, int priority);
 void process_destroy(Process *p);
-bool process_add_burst(Process *p, int cpu_time, int io_time);
+bool process_add_burst(Process *p, int64_t cpu_time, int64_t io_time);
 
 #endif // PROCESS_H
